panel/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt
argon2-cffi==21.2.0
python-dotenv==0.21.0
PyJWT==2.6.0
django-extensions==3.2.1
gunicorn==20.1.0
django-cors-headers==3.13.0
uvloop==0.21.0
httptools==0.6.4
coverage==6.4.4
requests==2.28.1
whitenoise
django-filter
urllib3
django-filter
django-allauth
celery==5.4.0
redis==3.5.3
telebot